Restructure core.schains.firewall package
